+++ src/Api/Service/AliasGroupSetter.php
@@ -44,7 +44,7 @@ class AliasGroupSetter {
 		);
 
 		$params['language'] = $aliasGroup->getLanguageCode();
-		$params['set'] = $aliasGroup->getAliases();
+		$params['set'] = implode( '|', $aliasGroup->getAliases() );
 
 		$params['token'] = $this->api->getToken();
 		$this->api->postRequest( new SimpleRequest( 'wbsetaliases', $params ) );